Work with ESP8266 2.4 lib

+++ build_all/edit_sample.ps1
@@ -10,14 +10,17 @@ param([string]$file,
 
 $ErrorActionPreference = "Stop"
 
-
-# Edit the sample iot_configs.h to have proper WiFi and subscription info
-Get-Content $file |
- % { $_ -replace “<Your WiFi network SSID or name>”, $ssid } |
- % { $_ -replace “<Your WiFi network WPA password or WEP key>”, $password } |
- % { $_ -replace “<host_name>”, $hostname } |
- % { $_ -replace “<device_id>”, $id } |
- % { $_ -replace “<device_key>”, $key }  -OutVariable content | Out-Null
- 
-$content | Set-Content $file
+# Ignore non-existent files
+if (Test-Path $file -PathType Leaf)
+{
+    # Edit the sample iot_configs.h to have proper WiFi and subscription info
+    Get-Content $file |
+     % { $_ -replace “<Your WiFi network SSID or name>”, $ssid } |
+     % { $_ -replace “<Your WiFi network WPA password or WEP key>”, $password } |
+     % { $_ -replace “<host_name>”, $hostname } |
+     % { $_ -replace “<device_id>”, $id } |
+     % { $_ -replace “<device_key>”, $key }  -OutVariable content | Out-Null
+     
+    $content | Set-Content $file
+}
